// types/woocommerce.d.ts
export interface VariationAttribute {
  id: number;
  name: string;
  option: string;
}